clean up code and use collections routes
