--- src/constants/PROFESSORSHIPS_MAP.js
@@ -1,119 +1,0 @@
-const PROFESSORSHIPS_MAP = {
-  'BASS UNIVERSITY FELLOWS IN UNDERGRADUATE EDUCATION PROGRAM': {
-    label: 'Bass University Fellows in Undergraduate Education Program',
-    section: 'Centers, Institutes, and Programs',
-    to: '/bass-university-fellows-in-undergraduate-education-program'
-  },
-  'CENTER FOR ADVANCED STUDY IN THE BEHAVIORAL SCIENCES': {
-    label: 'Center for Advanced Study in the Behavioral Sciences',
-    section: 'Centers, Institutes, and Programs',
-    to: '/center-for-advanced-study-in-the-behavioral-sciences'
-  },
-  'DEPARTMENT OF ATHLETICS, PHYSICAL EDUCATION, AND RECREATION': {
-    label: 'Department of Athletics, Physical Education, and Recreation',
-    section: 'Centers, Institutes, and Programs',
-    to: '/department-of-athletics-physical-education-and-recreation'
-  },
-  'FREEMAN SPOGLI INSTITUTE FOR INTERNATIONAL STUDIES': {
-    label: 'Freeman Spogli Institute for International Studies',
-    section: 'Centers, Institutes, and Programs',
-    to: '/freeman-spogli-institute-for-international-studies'
-  },
-  'GRADUATE SCHOOL OF BUSINESS': {
-    label: 'Graduate School of Business',
-    section: 'Schools',
-    to: '/graduate-school-of-business'
-  },
-  'GRADUATE SCHOOL OF EDUCATION': {
-    label: 'Graduate School of Education',
-    section: 'Schools',
-    to: '/graduate-school-of-education'
-  },
-  'HAAS CENTER FOR PUBLIC SERVICE': {
-    label: 'Haas Center for Public Service',
-    section: 'Centers, Institutes, and Programs',
-    to: '/haas-center-for-public-service'
-  },
-  'HOOVER INSTITUTION': {
-    label: 'Hoover Institution',
-    section: 'Centers, Institutes, and Programs',
-    to: '/hoover-institution'
-  },
-  'IRIS AND B. GERALD CANTOR CENTER FOR VISUAL ARTS': {
-    label: 'Iris and B. Gerald Cantor Center for Visual Arts',
-    section: 'Centers, Institutes, and Programs',
-    to: '/iris-and-b-gerald-cantor-center-for-visual-arts'
-  },
-  'SCHOOL OF EARTH, ENERGY AND ENVIRONMENTAL SCIENCES': {
-    label: 'School of Earth, Energy & Environmental Sciences',
-    section: 'Schools',
-    to: '/school-of-earth-energy-and-environmental-sciences'
-  },
-  'SCHOOL OF ENGINEERING': {
-    label: 'School of Engineering',
-    section: 'Schools',
-    to: '/school-of-engineering'
-  },
-  'SCHOOL OF HUMANITIES AND SCIENCES': {
-    label: 'School of Humanities and Sciences',
-    section: 'Schools',
-    to: '/school-of-humanities-and-sciences'
-  },
-  'STANFORD BIO-X': {
-    label: 'Stanford Bio-X',
-    section: 'Centers, Institutes, and Programs',
-    to: '/stanford-bio-x'
-  },
-  'STANFORD CHEM-H': {
-    label: 'Stanford ChEM-H',
-    section: 'Centers, Institutes, and Programs',
-    to: '/stanford-chem-h'
-  },
-  'STANFORD INSTITUTE FOR ECONOMIC POLICY RESEARCH': {
-    label: 'Stanford Institute for Economic Policy Research',
-    section: 'Centers, Institutes, and Programs',
-    to: '/stanford-institute-for-economic-policy-research'
-  },
-  'STANFORD INSTITUTE FOR HUMAN-CENTERED ARTIFICIAL INTELLIGENCE': {
-    label: 'Stanford Institute For Human-Centered Artificial Intelligence',
-    section: 'Centers, Institutes, and Programs',
-    to: '/stanford-institute-for-human-centered-artificial-intelligence'
-  },
-  'STANFORD KING CENTER ON GLOBAL DEVELOPMENT': {
-    label: 'Stanford King Center on Global Development',
-    section: 'Centers, Institutes, and Programs',
-    to: '/stanford-king-center-on-global-development'
-  },
-  'STANFORD LAW SCHOOL': {
-    label: 'Stanford Law School',
-    section: 'Schools',
-    to: '/stanford-law-school'
-  },
-  'STANFORD MEDICINE': {
-    label: 'Stanford Medicine',
-    section: 'Schools',
-    to: '/stanford-medicine'
-  },
-  'STANFORD UNIVERSITY LIBRARIES': {
-    label: 'Stanford Libraries',
-    section: 'Centers, Institutes, and Programs',
-    to: '/stanford-university-libraries'
-  },
-  'STANFORD WOODS INSTITUTE FOR THE ENVIRONMENT': {
-    label: 'Stanford Woods Institute for the Environment',
-    section: 'Centers, Institutes, and Programs',
-    to: 'stanford-woods-institute-for-the-environment'
-  },
-  'UNIVERSITY, PRESIDENTIAL, AND PROVOSTIAL PROFESSORSHIPS': {
-    label: 'University, Presidential, and Provostial Professorships',
-    section: 'Centers, Institutes, and Programs',
-    to: 'stanford-woods-institute-for-the-environment'
-  },
-  'WU TSAI NEUROSCIENCES INSTITUTE': {
-    label: 'Wu Tsai Neurosciences Institute',
-    section: 'Centers, Institutes, and Programs',
-    to: '/wu-tsai-neurosciences-institute'
-  }
-};
-
-export default PROFESSORSHIPS_MAP;
