+++ src/constants/professorships.json
@@ -1,0 +1,140 @@
+[
+  {
+    "id": "BASS UNIVERSITY FELLOWS IN UNDERGRADUATE EDUCATION PROGRAM",
+    "label": "Bass University Fellows in Undergraduate Education Program",
+    "section": "Centers, Institutes, and Programs",
+    "to": "bass-university-fellows-in-undergraduate-education-program"
+  },
+  {
+    "id": "CENTER FOR ADVANCED STUDY IN THE BEHAVIORAL SCIENCES",
+    "label": "Center for Advanced Study in the Behavioral Sciences",
+    "section": "Centers, Institutes, and Programs",
+    "to": "center-for-advanced-study-in-the-behavioral-sciences"
+  },
+  {
+    "id": "DEPARTMENT OF ATHLETICS, PHYSICAL EDUCATION, AND RECREATION",
+    "label": "Department of Athletics, Physical Education, and Recreation",
+    "section": "Centers, Institutes, and Programs",
+    "to": "department-of-athletics-physical-education-and-recreation"
+  },
+  {
+    "id": "FREEMAN SPOGLI INSTITUTE FOR INTERNATIONAL STUDIES",
+    "label": "Freeman Spogli Institute for International Studies",
+    "section": "Centers, Institutes, and Programs",
+    "to": "freeman-spogli-institute-for-international-studies"
+  },
+  {
+    "id": "GRADUATE SCHOOL OF BUSINESS",
+    "label": "Graduate School of Business",
+    "section": "Schools",
+    "to": "graduate-school-of-business"
+  },
+  {
+    "id": "GRADUATE SCHOOL OF EDUCATION",
+    "label": "Graduate School of Education",
+    "section": "Schools",
+    "to": "graduate-school-of-education"
+  },
+  {
+    "id": "HAAS CENTER FOR PUBLIC SERVICE",
+    "label": "Haas Center for Public Service",
+    "section": "Centers, Institutes, and Programs",
+    "to": "haas-center-for-public-service"
+  },
+  {
+    "id": "HOOVER INSTITUTION",
+    "label": "Hoover Institution",
+    "section": "Centers, Institutes, and Programs",
+    "to": "hoover-institution"
+  },
+  {
+    "id": "IRIS AND B. GERALD CANTOR CENTER FOR VISUAL ARTS",
+    "label": "Iris and B. Gerald Cantor Center for Visual Arts",
+    "section": "Centers, Institutes, and Programs",
+    "to": "iris-and-b-gerald-cantor-center-for-visual-arts"
+  },
+  {
+    "id": "SCHOOL OF EARTH, ENERGY AND ENVIRONMENTAL SCIENCES",
+    "label": "School of Earth, Energy & Environmental Sciences",
+    "section": "Schools",
+    "to": "school-of-earth-energy-and-environmental-sciences"
+  },
+  {
+    "id": "SCHOOL OF ENGINEERING",
+    "label": "School of Engineering",
+    "section": "Schools",
+    "to": "school-of-engineering"
+  },
+  {
+    "id": "SCHOOL OF HUMANITIES AND SCIENCES",
+    "label": "School of Humanities and Sciences",
+    "section": "Schools",
+    "to": "school-of-humanities-and-sciences"
+  },
+  {
+    "id": "STANFORD BIO-X",
+    "label": "Stanford Bio-X",
+    "section": "Centers, Institutes, and Programs",
+    "to": "stanford-bio-x"
+  },
+  {
+    "id": "STANFORD CHEM-H",
+    "label": "Stanford ChEM-H",
+    "section": "Centers, Institutes, and Programs",
+    "to": "stanford-chem-h"
+  },
+  {
+    "id": "STANFORD INSTITUTE FOR ECONOMIC POLICY RESEARCH",
+    "label": "Stanford Institute for Economic Policy Research",
+    "section": "Centers, Institutes, and Programs",
+    "to": "stanford-institute-for-economic-policy-research"
+  },
+  {
+    "id": "STANFORD INSTITUTE FOR HUMAN-CENTERED ARTIFICIAL INTELLIGENCE",
+    "label": "Stanford Institute For Human-Centered Artificial Intelligence",
+    "section": "Centers, Institutes, and Programs",
+    "to": "stanford-institute-for-human-centered-artificial-intelligence"
+  },
+  {
+    "id": "STANFORD KING CENTER ON GLOBAL DEVELOPMENT",
+    "label": "Stanford King Center on Global Development",
+    "section": "Centers, Institutes, and Programs",
+    "to": "stanford-king-center-on-global-development"
+  },
+  {
+    "id": "STANFORD LAW SCHOOL",
+    "label": "Stanford Law School",
+    "section": "Schools",
+    "to": "stanford-law-school"
+  },
+  {
+    "id": "STANFORD MEDICINE",
+    "label": "Stanford Medicine",
+    "section": "Schools",
+    "to": "stanford-medicine"
+  },
+  {
+    "id": "STANFORD UNIVERSITY LIBRARIES",
+    "label": "Stanford Libraries",
+    "section": "Centers, Institutes, and Programs",
+    "to": "stanford-university-libraries"
+  },
+  {
+    "id": "STANFORD WOODS INSTITUTE FOR THE ENVIRONMENT",
+    "label": "Stanford Woods Institute for the Environment",
+    "section": "Centers, Institutes, and Programs",
+    "to": "stanford-woods-institute-for-the-environment"
+  },
+  {
+    "id": "UNIVERSITY, PRESIDENTIAL, AND PROVOSTIAL PROFESSORSHIPS",
+    "label": "University, Presidential, and Provostial Professorships",
+    "section": "Centers, Institutes, and Programs",
+    "to": "stanford-woods-institute-for-the-environment"
+  },
+  {
+    "id": "WU TSAI NEUROSCIENCES INSTITUTE",
+    "label": "Wu Tsai Neurosciences Institute",
+    "section": "Centers, Institutes, and Programs",
+    "to": "wu-tsai-neurosciences-institute"
+  }
+]
