+++ src/pages/professorships/index.js
@@ -1,10 +1,8 @@
-import React, { useMemo, useState } from 'react';
-import { Link } from "gatsby"
+import React, { useState } from 'react';
+import { graphql, Link, useStaticQuery } from "gatsby"
 import Fuse from 'fuse.js';
-import { useTable } from 'react-table';
 
 import PROFESSORSHIPS from '../../fixtures/professorships.json';
-import PROFESSORSHIPS_MAP from '../../constants/PROFESSORSHIPS_MAP';
 
 const fuse = new Fuse(PROFESSORSHIPS, {
   keys: ['SUBCATEGORY', 'POSITION', 'CURRENT HOLDER'/*, 'WEBSITE'*/],
@@ -18,41 +16,38 @@ const getSubcategories = () => {
   return [...new Set(SUBS)];
 };
 
-const getTableDataBySubcategory = (subcategory) =>
-PROFESSORSHIPS.filter(item => item['SUBCATEGORY'] === subcategory);
+const convertArrayToObject = (array) => {
+  return array.reduce((obj, item) => (
+    {
+      ...obj,
+      [item?.node?.jsonId]: {
+        label: item?.node?.label,
+        section: item?.node?.section,
+        to: item?.node?.to
+      },
+    }
+  ), {});
+};
 
 
-console.log(fuse.search('business'));
-console.log(getSubcategories());
-
-export default function Index() {
-  const [getCategory, setCategory] = useState('BASS UNIVERSITY FELLOWS IN UNDERGRADUATE EDUCATION PROGRAM');
+const Index = () => {
   const [getSearchResults, setSearchResults] = useState(null);
   const [getSearchTerm, setSearchTerm] = useState('');
-  const columns = useMemo(() => [
-    {
-      Header: "Position",
-      accessor: "POSITION",
-    },
-    {
-      Header: "Current Holder",
-      accessor: "CURRENT HOLDER",
-    },
-  ], []);
-  const data = useMemo(() => getTableDataBySubcategory(getCategory).map((item, index) => {
-    return {
-      'POSITION': item['POSITION'],
-      'CURRENT HOLDER': item['CURRENT HOLDER'],
-      id: index
+  const {allProfessorshipsJson: { edges }} = useStaticQuery(graphql`
+    query {
+      allProfessorshipsJson {
+        edges {
+          node {
+            jsonId
+            label
+            to
+          }
+        }
+      }
     }
-  }, []));
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    rows,
-    prepareRow,
-  } = useTable({ columns, data });
+  `);
+
+  const professorshipsMap = convertArrayToObject(edges);
 
   return (
     <>
@@ -75,9 +70,9 @@ export default function Index() {
         </button>
       </div>
       {getSearchResults &&
-        getSearchResults.map(item => {
+        getSearchResults.map((item, index) => {
           return (
-            <dl>
+            <dl key={`${item.item['CURRENT HOLDER']}-${index}`}>
               <dt>{item.item['CURRENT HOLDER']}</dt>
               <dd>{item.item['POSITION']}</dd>
               <dd>{item.item['WEBSITE']}</dd>
@@ -88,60 +83,20 @@ export default function Index() {
       {!getSearchResults &&
         <>
           <ul css={{listStyle: 'none', margin: '0', padding: 0, display: 'flex', fontSize: '12px', flexWrap: 'wrap'}}>
-            {getSubcategories().map(item => (
-              <li
-                css={{margin: '10px', cursor: 'pointer'}}
-                onClick={(event) => setCategory(event.target.innerHTML)}
-              >
-                <Link to={`/professorships${PROFESSORSHIPS_MAP[item].to}`}>
-                  {PROFESSORSHIPS_MAP[item].label}
-                </Link>
-              </li>
-            ))}
+            {getSubcategories().map(item => {
+              return (
+                <li css={{margin: '10px', cursor: 'pointer'}} key={item}>
+                  <Link to={`/professorships/${professorshipsMap[item].to}`}>
+                    {professorshipsMap[item].label}
+                  </Link>
+                </li>
+              );
+            })}
           </ul>
-          <table {...getTableProps()} css={{ border: 'solid 1px gray' }}>
-            <thead>
-              {headerGroups.map(headerGroup => (
-                <tr {...headerGroup.getHeaderGroupProps()}>
-                  {headerGroup.headers.map(column => (
-                    <th
-                      {...column.getHeaderProps()}
-                      css={{
-                        color: 'black',
-                        fontWeight: 'bold',
-                      }}
-                    >
-                      {column.render('Header')}
-                    </th>
-                  ))}
-                </tr>
-              ))}
-            </thead>
-            <tbody {...getTableBodyProps()}>
-              {rows.map(row => {
-                prepareRow(row)
-                return (
-                  <tr {...row.getRowProps()}>
-                    {row.cells.map(cell => {
-                      return (
-                        <td
-                          {...cell.getCellProps()}
-                          css={{
-                            padding: '10px',
-                            border: 'solid 1px gray',
-                          }}
-                        >
-                          {cell.render('Cell')}
-                        </td>
-                      )
-                    })}
-                  </tr>
-                )
-              })}
-            </tbody>
-          </table>
         </>
       }
     </>
   );
-}
+};
+
+export default Index;
